# docs/api_doc/swagger.json
{"openapi":"3.0.1","info":{"title":"'@project.name@'","description":"'@project.description@'","contact":{"name":"John Doe","email":"dumy@dumy"},"version":"'@project.version@'"},"servers":[{"url":"http://localhost:8080","description":"'@project.description@'"}],"paths":{"/visitors":{"get":{"tags":["visitor-controller"],"operationId":"getAllVisitors","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"lastName","in":"query","required":false,"schema":{"type":"string"}},{"name":"filter","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VisitorDTO"}}}}}}},"put":{"tags":["visitor-controller"],"operationId":"generateVisitor","parameters":[{"name":"visitorId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VisitorRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VisitorDTO"}}}}}}},"/auths":{"get":{"tags":["auth-controller"],"operationId":"getAuth","parameters":[{"name":"docNumber","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AuthDTO"}}}}}}},"post":{"tags":["auth-controller"],"operationId":"authorizeVisitor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VisitorAuthRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthDTO"}}}}}}},"/visitors/{visitorId}":{"get":{"tags":["visitor-controller"],"operationId":"getVisitorById","parameters":[{"name":"visitorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VisitorDTO"}}}}}},"delete":{"tags":["visitor-controller"],"operationId":"deleteVisitor","parameters":[{"name":"visitorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VisitorDTO"}}}}}}},"/visitors/by-doc-number/{docNumber}":{"get":{"tags":["visitor-controller"],"operationId":"getVisitorByDocNumber","parameters":[{"name":"docNumber","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VisitorDTO"}}}}}}},"/qr/{docNumber}":{"get":{"tags":["qr-controller"],"operationId":"generateQr","parameters":[{"name":"docNumber","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/auths/valids":{"get":{"tags":["auth-controller"],"operationId":"getValidAuths","parameters":[{"name":"docNumber","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AuthDTO"}}}}}}}},"/auth-ranges":{"get":{"tags":["auth-range-controller"],"operationId":"getAuthRanges","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AuthRangeDTO"}}}}}}}}},"components":{"schemas":{"VisitorRequest":{"type":"object","properties":{"name":{"type":"string"},"last_name":{"type":"string"},"doc_type":{"type":"string","enum":["DNI","PASSPORT","CUIL","CUIT"]},"doc_number":{"type":"integer","format":"int64"},"birth_date":{"pattern":"dd-MM-yyyy","type":"string","example":"18-09-2024"},"is_active":{"type":"boolean"}}},"VisitorDTO":{"type":"object","properties":{"visitor_id":{"type":"integer","format":"int64"},"name":{"type":"string"},"last_name":{"type":"string"},"doc_type":{"type":"string","enum":["DNI","PASSPORT","CUIL","CUIT"]},"doc_number":{"type":"integer","format":"int64"},"birth_date":{"type":"string","format":"date"},"is_active":{"type":"boolean"}}},"AuthRangeRequestDTO":{"type":"object","properties":{"date_from":{"pattern":"dd-MM-yyyy","type":"string","description":"The start date of the authorized range.","example":"01-01-2022"},"date_to":{"pattern":"dd-MM-yyyy","type":"string","description":"The end date of the authorized range.","example":"02-01-2024"},"hour_from":{"pattern":"HH:mm:ss","type":"string","description":"The start time of the authorized range.","example":"08:00:00"},"hour_to":{"pattern":"HH:mm:ss","type":"string","description":"The end time of the authorized range.","example":"17:00:00"},"days_of_week":{"type":"array","items":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]}},"plot_id":{"type":"integer","format":"int64"},"comment":{"type":"string"},"is_active":{"type":"boolean"}}},"VisitorAuthRequest":{"type":"object","properties":{"visitorType":{"type":"string","enum":["OWNER","WORKER","VISITOR","EMPLOYEE","SUPPLIER"]},"visitorRequest":{"$ref":"#/components/schemas/VisitorRequest"},"authRangeRequest":{"type":"array","items":{"$ref":"#/components/schemas/AuthRangeRequestDTO"}}}},"AuthDTO":{"type":"object","properties":{"auth_id":{"type":"integer","format":"int64"},"visitor":{"$ref":"#/components/schemas/VisitorDTO"},"visitor_type":{"type":"string","enum":["OWNER","WORKER","VISITOR","EMPLOYEE","SUPPLIER"]},"auth_ranges":{"type":"array","items":{"$ref":"#/components/schemas/AuthRangeDTO"}},"is_active":{"type":"boolean"}}},"AuthRangeDTO":{"type":"object","properties":{"auth_range_id":{"type":"integer","format":"int64"},"date_from":{"pattern":"dd-MM-yyyy","type":"string","description":"The start date of the authorized range.","example":"01-01-2022"},"date_to":{"pattern":"dd-MM-yyyy","type":"string","description":"The end date of the authorized range.","example":"02-01-2024"},"hour_from":{"pattern":"HH:mm:ss","type":"string","description":"The start time of the authorized range.","example":"08:00:00"},"hour_to":{"pattern":"HH:mm:ss","type":"string","description":"The end time of the authorized range.","example":"17:00:00"},"days_of_week":{"type":"array","items":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]}},"plot_id":{"type":"integer","format":"int64"},"comment":{"type":"string"},"is_active":{"type":"boolean"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}